feat: upsert User when reading session info
