+++ src/app/(withAuth)/courses/page.tsx
@@ -1,8 +1,13 @@
 import type { NextPage } from 'next';
 
+import { getSessionOnServerPage } from '../../../infrastructures/session';
+
 const CoursePage: NextPage = async () => {
+  const session = await getSessionOnServerPage();
+
   return (
     <main>
+      <div>UserID on Supertokens: {session.getUserId()}</div>
       <h1>Courses</h1>
     </main>
   );
