+++ src/infrastructures/supertokens/backendConfig.ts
@@ -20,7 +20,10 @@ export const backendConfig = (): TypeInput => {
 };
 
 let initialized = false;
-// This function is used in your APIs to make sure SuperTokens is initialised
+
+/**
+ * This function is used in your APIs to make sure SuperTokens is initialised
+ */
 export function ensureSuperTokensInit(): void {
   if (!initialized) {
     SuperTokens.init(backendConfig());
